# video/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /douyin/favorite/action:
        post:
            tags:
                - FavoriteService
            description: |-
                点赞/取消点赞
                 登录用户才能点赞
            operationId: FavoriteService_Favorite
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/favorite.api.favorite.v1.Douyin_favorite_action_request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.Douyin_favorite_action_response'
    /douyin/favorite/count/user:
        get:
            tags:
                - FavoriteService
            description: 获取用户的获赞数和点赞数量
            operationId: FavoriteService_GetFavoriteCntByUId
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.GetFavoriteCntByUIdResponse'
    /douyin/favorite/count/video:
        get:
            tags:
                - FavoriteService
            description: 获取视频的获赞数
            operationId: FavoriteService_GetFavoriteCntByVId
            parameters:
                - name: id
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.GetFavoriteCntByVIdResponse'
    /douyin/favorite/is_favorite:
        get:
            tags:
                - FavoriteService
            description: 是否喜欢视频
            operationId: FavoriteService_GetIsFavorite
            parameters:
                - name: videoId
                  in: query
                  schema:
                    type: string
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.GetIsFavoriteResponse'
    /douyin/favorite/list:
        get:
            tags:
                - FavoriteService
            description: 获取用户点赞视频列表
            operationId: FavoriteService_FavoriteList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: token
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/favorite.api.favorite.v1.Douyin_favorite_list_response'
        post:
            tags:
                - VideoService
            description: 通过喜欢视频 id 列表获取视频列表
            operationId: VideoService_FavoriteListByVId
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/video.api.video.v1.FavoriteListReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/video.api.video.v1.FavoriteListResp'
...
    /douyin/message/action:
        post:
            tags:
                - MessageService
            description: 发送消息
            operationId: MessageService_MessageSend
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/message.api.message.v1.Douyin_message_send_request'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message.api.message.v1.Douyin_message_send_response'
    /douyin/message/chat:
        get:
            tags:
                - MessageService
            description: 获取历史消息记录
            operationId: MessageService_MessageRecord
            parameters:
                - name: token
                  in: query
                  schema:
                    type: string
                - name: toUserId
                  in: query
                  schema:
                    type: string
                - name: preMsgTime
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message.api.message.v1.Douyin_message_record_response'
    /douyin/message/latest:
        get:
            tags:
                - MessageService
            description: 获取用户和每个好友的最新一条消息
            operationId: MessageService_GetNewMessages
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: toUserId
                  in: query
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message.api.message.v1.GetNewMessages_response'
...
components:
    schemas:
...
        favorite.api.favorite.v1.Douyin_favorite_action_request:
            type: object
            properties:
                token:
                    type: string
                videoId:
                    type: string
                actionType:
                    type: integer
                    format: int32
        favorite.api.favorite.v1.Douyin_favorite_action_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        favorite.api.favorite.v1.Douyin_favorite_list_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/favorite.api.favorite.v1.Video'
        favorite.api.favorite.v1.GetFavoriteCntByUIdResponse:
            type: object
            properties:
                totalFavorited:
                    type: string
                favoriteCount:
                    type: string
        favorite.api.favorite.v1.GetFavoriteCntByVIdResponse:
            type: object
            properties:
                favoriteCount:
                    type: string
        favorite.api.favorite.v1.GetIsFavoriteResponse:
            type: object
            properties:
                Favorite:
                    type: boolean
        favorite.api.favorite.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
        favorite.api.favorite.v1.Video:
            type: object
            properties:
                id:
                    type: string
                author:
                    $ref: '#/components/schemas/favorite.api.favorite.v1.User'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                commentCount:
                    type: string
                isFavorite:
                    type: boolean
                title:
                    type: string
        message.api.message.v1.Douyin_message_record_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
                messageList:
                    type: array
                    items:
                        $ref: '#/components/schemas/message.api.message.v1.Message'
        message.api.message.v1.Douyin_message_send_request:
            type: object
            properties:
                token:
                    type: string
                toUserId:
                    type: string
                sendType:
                    type: integer
                    format: int32
                content:
                    type: string
        message.api.message.v1.Douyin_message_send_response:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                statusMsg:
                    type: string
        message.api.message.v1.GetNewMessages_response:
            type: object
            properties:
                latestMessageList:
                    type: array
                    items:
                        $ref: '#/components/schemas/message.api.message.v1.LatestMessage'
        message.api.message.v1.LatestMessage:
            type: object
            properties:
                content:
                    type: string
                msgType:
                    type: string
                friendId:
                    type: string
        message.api.message.v1.Message:
            type: object
            properties:
                id:
                    type: string
                toUserId:
                    type: string
                fromUserId:
                    type: string
                content:
                    type: string
                createTime:
                    type: string
...
        video.api.video.v1.FavoriteListReq:
            type: object
            properties:
                videoIdList:
                    type: array
                    items:
                        type: string
        video.api.video.v1.FavoriteListResp:
            type: object
            properties:
                videoList:
                    type: array
                    items:
                        $ref: '#/components/schemas/video.api.video.v1.Video'
...
        video.api.video.v1.User:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                followCount:
                    type: string
                followerCount:
                    type: string
                isFollow:
                    type: boolean
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                totalFavorited:
                    type: string
                workCount:
                    type: string
                favoriteCount:
                    type: string
        video.api.video.v1.Video:
            type: object
            properties:
                id:
                    type: string
                author:
                    $ref: '#/components/schemas/video.api.video.v1.User'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                commentCount:
                    type: string
                isFavorite:
                    type: boolean
                title:
                    type: string
...